+++ custom_components/octopus_energy_de/coordinator.py
@@ -7,8 +7,9 @@ from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, Upda
 
 from .api.client import OctopusEnergyDEClient
 from .api.exceptions import OctopusEnergyDEError
+from .api.models.smartflex import SmartFlexSnapshot
 from .api.models.tariff import AccountSnapshot
-from .const import DEFAULT_SCAN_INTERVAL, DOMAIN
+from .const import BASE_SCAN_INTERVAL, DOMAIN, METER_SCAN_INTERVAL, SMARTFLEX_SCAN_INTERVAL
 
 
 class OctopusEnergyDECoordinator(DataUpdateCoordinator[AccountSnapshot]):
@@ -17,7 +18,7 @@ class OctopusEnergyDECoordinator(DataUpdateCoordinator[AccountSnapshot]):
             hass,
             logger=__import__("logging").getLogger(__name__),
             name=f"{DOMAIN}_{account_number}",
-            update_interval=DEFAULT_SCAN_INTERVAL,
+            update_interval=BASE_SCAN_INTERVAL,
         )
         self.client = client
         self.account_number = account_number
@@ -27,3 +28,50 @@ class OctopusEnergyDECoordinator(DataUpdateCoordinator[AccountSnapshot]):
             return await self.client.account_snapshot(self.account_number)
         except OctopusEnergyDEError as err:
             raise UpdateFailed(str(err)) from err
+
+
+class OctopusEnergyDEMeterCoordinator(DataUpdateCoordinator[AccountSnapshot]):
+    def __init__(
+        self,
+        hass: HomeAssistant,
+        client: OctopusEnergyDEClient,
+        account_number: str,
+        base_coordinator: OctopusEnergyDECoordinator,
+    ) -> None:
+        super().__init__(
+            hass,
+            logger=__import__("logging").getLogger(__name__),
+            name=f"{DOMAIN}_{account_number}_meters",
+            update_interval=METER_SCAN_INTERVAL,
+        )
+        self.client = client
+        self.account_number = account_number
+        self.base_coordinator = base_coordinator
+
+    async def _async_update_data(self) -> AccountSnapshot:
+        try:
+            return await self.client.meter_snapshot(
+                self.account_number, self.base_coordinator.data.electricity
+            )
+        except OctopusEnergyDEError as err:
+            raise UpdateFailed(str(err)) from err
+
+
+class OctopusEnergyDESmartFlexCoordinator(DataUpdateCoordinator[SmartFlexSnapshot]):
+    def __init__(self, hass: HomeAssistant, client: OctopusEnergyDEClient, account_number: str) -> None:
+        super().__init__(
+            hass,
+            logger=__import__("logging").getLogger(__name__),
+            name=f"{DOMAIN}_{account_number}_smartflex",
+            update_interval=SMARTFLEX_SCAN_INTERVAL,
+        )
+        self.client = client
+        self.account_number = account_number
+
+    async def _async_update_data(self) -> SmartFlexSnapshot:
+        try:
+            return await self.client.smartflex_snapshot(self.account_number)
+        except OctopusEnergyDEError as err:
+            self.logger.debug(
+                "SmartFlex data unavailable for account: %s", err)
+            return SmartFlexSnapshot()